idle: exit the cpu_idle_poll loop if cpu_idle_force_poll is cleared
cpu_idle_poll_ctrl allows the enabling/disabling of the idle
polling mode; this mode allows a CPU to spin waiting for a
new task to be scheduled rather than having to execute the
arch specific idle code.

However, the loop that checks for a new task does not look
at the flag that enables idle polling mode. So, the CPU may
continue to spin even though the aforementioned flag has
been cleared. Since the CPU is already in idle, it may be
a while before a task is scheduled, precluding potential
power savings.

Modify the while loop conditional in question to also check
if the cpu_idle_force_poll flag is set.

Change-Id: If4752a39acc087cc1c4c0389ab72b050638c54f6
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>